Fix merge artifacts once again

+++ Minigames/src/main/java/au/com/mineauz/minigames/events/MinigameTimerTickEvent.java
@@ -12,7 +12,7 @@ public class MinigameTimerTickEvent extends AbstractCancellableMinigameEvent {
         this.timer = timer;
     }
 
-    public int getTimeLeft() {
+    public long getTimeLeft() {
         return timer.getTimeLeft();
     }
